feat: Added integration tests for IPFS/Full State Recovery (#299)
This PR implements integration tests for IPFS + full state recovery by killing the worker and respawning it in order to trigger the recovery from the storage. It tests for balances and leaves the nonce + orderbook states to be tested by the rest of the operations, because if the state fails to recover the orderbook based operations will fail due to mismatching nonce and missing orders. 
Closes #291